string.c: String dimensions were declared `short`, which overflows easily, and was not caught by the logic. This fix makes them bigger (effectively infinite, modulo system RAM).
Fixes #3

+++ edam.h
@@ -59,8 +59,8 @@ typedef struct Disc{
    }block;         /* array of used blocks (same shape as List) */
 }Disc;
 typedef struct String{
-   short   n;
-   short   size;
+   ulong   n;
+   ulong   size;
    char   *s;
 }String;
 typedef struct Buffer{
